Add simple integration test based on cp(1).
This test distills some of the essential elements of cp(1)'s utils.c
that are relevant to intraprocedural information flow. It currently only
tests those bits that we have finished, i.e., adapting IF sources but
not sinks, so it should pass.

The expectations expressed in the `PROVCHECK` lines of this test should,
over time, start to look like the handcrafted `cp_metaio(1)` code
attached to Issue cadets/freebsd#58.