Make FlowFinder use MemorySSA.
MemorySSA provides a mechanism for explicitly reasoning about memory versioning, and with a MemorySSAWalker we can intelligently answer questions such as, "which (set of) store instruction(s) most recently clobbered the location we are currently reading from?" Switch FlowFinder to use this mechanism instead of the previous ad hoc code that essentially tried to achieve the same results but without proper alias analysis, etc.

This new approach produces a much more complete graph of pairwise information flows from within a function. Unfortunately it also exposes a prevously-less-exposed issue: information flow cycles in which an instruction may depend on a previous invocation of that same instruction in a loop. For example:

```c
while ((n = read(fd, buffer + n, sizeof(buffer) - n)) > 0)
	total += n;
```

Add an XFAIL test to describe this problem.